requirements.txt
numpy==1.24.3
Pillow==9.4.0
streamlit==1.25.0
tensorflow==2.13.0
tensorflow_macos==2.13.0
